+++ package.json
@@ -214,13 +214,13 @@
                 "icon": "$(panel-close)"
             },
             {
-                "command": "clangd.projectSettings",
-                "title": "clangd: Open Project Settings",
+                "command": "clangd.projectConfig",
+                "title": "clangd: Open project configuration file",
                 "icon": "$(gear)"
             },
             {
-                "command": "clangd.globalSettings",
-                "title": "clangd: Open Global Settings",
+                "command": "clangd.userConfig",
+                "title": "clangd: Open user configuration file",
                 "icon": "$(gear)"
             }
         ],
